refactor(oci): split runtime config
